pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sdlc-tools"
version = "0.1.0"
description = "A developer CLI toolkit for SDLC automation — AI code impact reports, release tagging, and PR management."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "White-Yaksha" }]
keywords = ["sdlc", "ci", "cd", "automation", "github", "cli", "devtools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "responses>=0.24",
    "ruff>=0.3",
    "coverage>=7.4",
]

[project.scripts]
sdlc-tools = "sdlc_tools.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sdlc_tools"]

[tool.hatch.build.targets.wheel.force-include]
"src/sdlc_tools/prompts/default.txt" = "sdlc_tools/prompts/default.txt"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
